:sparkles: Add Dynamic output, change logic of pipeline and use new IoManagers

+++ src/solids/iiif.py
@@ -1,31 +1,26 @@
-import json
-
+import io
 import os
-import shutil
 import subprocess
 from math import *
-from typing import Collection, List
+from typing import Collection, List, Mapping
 
-import boto3
-from botocore import endpoint
 import dagster as dg
 import pandas as pd
 import requests
-from dagster.builtins import Nothing, String
 from dotenv import load_dotenv
-from requests.api import get
 from IIIFpres import iiifpapi3
 from IIIFpres.utilities import *
+from PIL import Image
 from requests.adapters import HTTPAdapter
+from requests.api import get
 from urllib3.util import Retry
-from tqdm import tqdm
+
 from solids.export import *
 
 load_dotenv(override=True)
 
 iiifpapi3.BASE_URL = "https://imaginerio-images.s3.us-east-1.amazonaws.com/"
 iiifpapi3.LANGUAGES = ["pt-BR", "en"]
-# iiifpapi3.BASE_URL = "http://127.0.0.1:8000/"
 
 
 def info_json_exists(identifier):
@@ -52,39 +47,59 @@ def info_json_exists(identifier):
         return False
 
 
-def tile_image(identifier):
+@dg.solid(output_defs=[dg.OutputDefinition(io_manager_key="iiif_manager")])
+def tile_image(context, item):
+    identifier = item["identifier"]
 
-    img_data = requests.get(
+    if not info_json_exists(identifier) or context.mode_def.name == "test":
+        img_data = requests.get(
+            "https://imaginerio-images.s3.us-east-1.amazonaws.com/iiif/{0}/full/max/0/default.jpg".format(
+                identifier
+            )
+        ).content
+
+        img_path = f"iiif/{identifier}.jpg"
+
+        if not os.path.exists("iiif"):
+            os.mkdir("iiif")
+
+        with open(img_path, "wb") as handler:
+            handler.write(img_data)
+
+        command = ["java", "-jar", "utils/mod-tiler.jar", img_path]
+
+        process = subprocess.Popen(
+            command,
+        )
+        process.communicate()
+        os.remove(img_path)
+
+        return [
+            {"data": "iiif/{0}".format(identifier), "type": "path"},
+        ]
+    else:
+        pass
+
+
+@dg.solid(
+    input_defs=[dg.InputDefinition("mapping", root_manager_key="mapping_root")],
+    output_defs=[dg.OutputDefinition(io_manager_key="iiif_manager")],
+)
+def write_manifests(context, item, mapping):
+    identifier = item["identifier"]
+    item = item["row"]
+    context.log.info("Processing: {0}".format(str(identifier)))
+
+    # Get image sizes
+    img = requests.get(
         "https://imaginerio-images.s3.us-east-1.amazonaws.com/iiif/{0}/full/max/0/default.jpg".format(
             identifier
         )
     ).content
+    image = Image.open(io.BytesIO(img))
+    imgwidth, imgheight = image.size
+    mapping.set_index("Label:en", inplace=True)
 
-    img_path = f"tmp/{identifier}.jpg"
-
-    if not os.path.exists("tmp"):
-        os.mkdir("tmp")
-
-    with open(img_path, "wb") as handler:
-        handler.write(img_data)
-
-    command = ["java", "-jar", "utils/mod-tiler.jar", img_path]
-
-    process = subprocess.Popen(
-        command,
-        # cwd="src"
-        # shell=True,
-        # stdout=subprocess.PIPE,
-        # stderr=subprocess.PIPE,
-    )
-    process.communicate()
-
-    os.remove(img_path)
-    info = json.load(open("tmp/iiif/{0}/info.json".format(identifier)))
-    return info
-
-
-def write_manifest(info, identifier, item, mapping):
     def get_multilingual_values(en):
         url = "http://wikidata.org/wiki/{0}".format(mapping.loc[en, "Wiki ID"])
         pt = mapping.loc[en, "Label:pt"]
@@ -258,8 +273,8 @@ def write_manifest(info, identifier, item, mapping):
 
     # Thumbnail
     thumbnail = iiifpapi3.thumbnail()
-    thumb_width = info["sizes"][0]["width"]
-    thumb_height = info["sizes"][0]["height"]
+    thumb_width = int(imgwidth / 16)
+    thumb_height = int(imgheight / 16)
     thumbnail.set_id(
         extendbase_url="iiif/{0}/full/{1},{2}/0/default.jpg".format(
             str(identifier).replace(" ", "_"), thumb_width, thumb_height
@@ -290,10 +305,6 @@ def write_manifest(info, identifier, item, mapping):
     item_provider.add_homepage(item_homepage)
     manifest.add_provider(item_provider)
 
-    # Get image sizes list
-    imgwidth = info["width"]
-    imgheight = info["height"]
-
     # Canvas
     canvas = manifest.add_canvas_to_items()
     canvas.set_id(extendbase_url="canvas/p1")
@@ -318,16 +329,15 @@ def write_manifest(info, identifier, item, mapping):
     s.set_type("ImageService3")
     s.set_profile("level0")
 
-    # Save manifest
-    if not os.path.exists("tmp/iiif/{0}".format(identifier)):
-        os.makedirs("tmp/iiif/{0}".format(identifier))
-    file = "tmp/iiif/{0}/manifest.json".format((identifier))
-    manifest.json_save(file)
-    return manifest
+    # setup items to send to io_managers
+    manifest_path = "iiif/{0}/manifest.json".format((identifier))
+    manifest_obj = manifest.json_dumps(
+        dumps_errors=False,
+        ensure_ascii=False,
+        context="http://iiif.io/api/presentation/3/context.json",
+    )
+    data = [{"data": manifest_obj, "key": manifest_path, "type": "json"}]
 
-
-def append_to_collection(manifest, collection_name):
-    collection_path = "tmp/iiif/collection/{0}.json".format(collection_name)
     # Logo
     logo = iiifpapi3.logo()
     logo.set_id(
@@ -335,120 +345,108 @@ def append_to_collection(manifest, collection_name):
     )
     logo.set_format("image/png")
     logo.set_hightwidth(164, 708)
-    try:
-        endpoint = "https://imaginerio-images.s3.us-east-1.amazonaws.com/iiif/collection/{0}.json".format(
-            collection_name
-        )
-        retry_strategy = Retry(
-            total=3,
-            status_forcelist=[429, 500, 502, 503, 504],
-            method_whitelist=["HEAD", "GET", "OPTIONS"],
-        )
-        adapter = HTTPAdapter(max_retries=retry_strategy)
-        http = requests.Session()
-        http.mount("https://", adapter)
-        http.mount("http://", adapter)
-        collection_data = http.get(endpoint).json()
-        collection = read_API3_json_dict(collection_data)
-    # TO-DO don't wait for json_save to throw exception, check for response code instead
-    except Exception as e:
-        print(e)
-        print("Couldn't find collection. Creating...")
-        # Homepage
-        collection_homepage = iiifpapi3.homepage()
-        collection_homepage.set_id("https://imaginerio.org")
-        collection_homepage.set_type("Text")
-        collection_homepage.add_label("none", "imagineRio")
-        collection_homepage.set_format("text/html")
 
-        # Provider
-        collection_provider = iiifpapi3.provider()
-        collection_provider.set_id("https://imaginerio.org")
-        collection_provider.add_label("en", "imagineRio")
-        collection_provider.add_label("pt-BR", "imagineRio")
-        collection_provider.add_logo(logo)
-        collection_provider.add_homepage(collection_homepage)
+    for collection_name in item["Item Set"].lower().split("||"):
+        collection_path = "iiif/collection/{0}.json".format(collection_name)
 
-        # Collection manifest
-        collection = iiifpapi3.Collection()
-        collection.set_id(extendbase_url="iiif/collection/{0}".format(collection_name))
-        collection.add_label("en", collection_name)
-        collection.add_label("pt-BR", collection_name)
-        collection.add_requiredStatement(
-            label="Attribution",
-            value="Hosted by imagineRio",
-            language_l="en",
-            language_v="en",
-        )
-        collection.add_requiredStatement(
-            label="Attribution",
-            value="Hospedado por imagineRio",
-            language_l="pt-BR",
-            language_v="pt-BR",
-        )
-        collection.set_rights("http://rightsstatements.org/vocab/CNE/1.0/")
+        try:
+            if context.mode_def.name == "test":
+                collection = read_API3_json(collection_path)
+            else:
+                endpoint = "https://imaginerio-images.s3.us-east-1.amazonaws.com/iiif/collection/{0}.json".format(
+                    collection_name
+                )
+                retry_strategy = Retry(
+                    total=3,
+                    status_forcelist=[429, 500, 502, 503, 504],
+                    method_whitelist=["HEAD", "GET", "OPTIONS"],
+                )
+                adapter = HTTPAdapter(max_retries=retry_strategy)
+                http = requests.Session()
+                http.mount("https://", adapter)
+                http.mount("http://", adapter)
+                collection_data = http.get(endpoint).json()
+                collection = read_API3_json_dict(collection_data)
 
-        thumbnailobj = iiifpapi3.thumbnail()
-        thumbnailobj.set_id(
-            extendbase_url="iiif/0071824cx001-01/full/144,95/0/default.jpg"
-        )
-        thumbnailobj.set_hightwidth(95, 144)
-        collection.add_thumbnail(thumbnailobj=thumbnailobj)
+            # Deal with duplicates
+            for item in collection.items:
+                if "{0}/manifest".format(identifier) in item.id:
+                    collection.items.remove(item)
 
-        collection.add_provider(collection_provider)
+        except Exception as e:
+            print(e)
+            print("Couldn't find collection: {0}. Creating...".format(collection_name))
+            # Homepage
+            collection_homepage = iiifpapi3.homepage()
+            collection_homepage.set_id("https://imaginerio.org")
+            collection_homepage.set_type("Text")
+            collection_homepage.add_label("none", "imagineRio")
+            collection_homepage.set_format("text/html")
 
-    if not os.path.exists("tmp/iiif/collection"):
-        os.mkdir("tmp/iiif/collection")
-    collection.add_manifest_to_items(manifest)
-    collection.json_save(collection_path)
-    print("Collection updated!")
+            # Provider
+            collection_provider = iiifpapi3.provider()
+            collection_provider.set_id("https://imaginerio.org")
+            collection_provider.add_label("en", "imagineRio")
+            collection_provider.add_label("pt-BR", "imagineRio")
+            collection_provider.add_logo(logo)
+            collection_provider.add_homepage(collection_homepage)
 
-
-def upload_to_cloud():
-    print("Uploading...")
-    S3 = boto3.client("s3")
-    BUCKET = "imaginerio-images"
-    upload_log = []
-    for root, dirs, files in os.walk("tmp"):
-        for file in files:
-            path = root.split("tmp/")[-1] + "/" + file
-            content = (
-                "image/jpeg" if file.split(".")[-1] == "jpg" else "application/json"
+            # Collection manifest
+            collection = iiifpapi3.Collection()
+            collection.set_id(
+                extendbase_url="iiif/collection/{0}".format(collection_name)
             )
-            try:
-                S3.upload_file(
-                    os.path.join(root, file),
-                    BUCKET,
-                    path,
-                    ExtraArgs={"ContentType": content},
-                )
-                print(
-                    os.path.join(root, file),
-                    BUCKET,
-                    path,
-                    content,
-                )
-            except Exception as e:
-                upload_log.append(path)
-                print("Couldn't upload image {0}.Error: {1}".format(path, e))
+            collection.add_label("en", collection_name)
+            collection.add_label("pt-BR", collection_name)
+            collection.add_requiredStatement(
+                label="Attribution",
+                value="Hosted by imagineRio",
+                language_l="en",
+                language_v="en",
+            )
+            collection.add_requiredStatement(
+                label="Attribution",
+                value="Hospedado por imagineRio",
+                language_l="pt-BR",
+                language_v="pt-BR",
+            )
+            collection.set_rights("http://rightsstatements.org/vocab/CNE/1.0/")
 
-    if not upload_log:
-        print("Cleaning tmp folder...")
-        shutil.rmtree("tmp")
+            thumbnailobj = iiifpapi3.thumbnail()
+            thumbnailobj.set_id(
+                extendbase_url="iiif/0071824cx001-01/full/144,95/0/default.jpg"
+            )
+            thumbnailobj.set_hightwidth(95, 144)
+            collection.add_thumbnail(thumbnailobj=thumbnailobj)
+            collection.add_provider(collection_provider)
+
+        collection.add_manifest_to_items(manifest)
+
+        # setup items to send to io_managers
+        collection = collection.json_dumps(
+            dumps_errors=False,
+            ensure_ascii=False,
+            context="http://iiif.io/api/presentation/3/context.json",
+        )
+        collections = {}
+        collections["data"] = collection
+        collections["key"] = collection_path
+        collections["type"] = "json"
+        data.append(collections)
+
+        print("Collection updated: {0}".format(collection_name))
+
+    return data
 
 
 @dg.solid(
-    # config_schema={
-    #     "debug": dg.BoolSource,
-    # },
     input_defs=[
         dg.InputDefinition("metadata", root_manager_key="metadata_root"),
-        dg.InputDefinition("mapping", root_manager_key="mapping_root"),
     ],
+    output_defs=[dg.DynamicOutputDefinition(dict)],
 )
-def iiify(context, metadata, mapping):
+def get_items(context, metadata):
     metadata.set_index("Source ID", inplace=True)
-    mapping.set_index("Label:en", inplace=True)
     metadata.dropna(
         subset=[
             "Source",
@@ -462,17 +460,9 @@ def iiify(context, metadata, mapping):
     )
     metadata = metadata.loc[metadata["Source"] == "Instituto Moreira Salles"]
     if context.mode_def.name == "test":
-        metadata = metadata[-2:]
+        metadata = metadata[:2]
     for identifier, item in metadata.iterrows():
-        if info_json_exists(identifier) or context.solid_config["manifest_only"]:
-            context.log.info("Skipping: {0}".format(str(identifier)))
-            continue
-        else:
-            context.log.info("Processing: {0}".format(str(identifier)))
-            info = tile_image(identifier)
-            manifest = write_manifest(info, identifier, item, mapping)
-            for collection_name in item["Item Set"].lower().split("||"):
-                append_to_collection(manifest, collection_name)
-            # if not context.solid_config["debug"]:
-            if context.mode_def.name == "prod":
-                upload_to_cloud()
+        yield dg.DynamicOutput(
+            value={"identifier": identifier, "row": item},
+            mapping_key=identifier.replace("-", "_"),
+        )
